+++ tests/rest.spec.ts
@@ -1,0 +1,252 @@
+// todo
+// tests commented out that have:
+//   if (!process.version.match(/^v12/)) return
+// because they fail in node v10 tests
+// either remove them, rewrite them to be agnostic, or find way to run them only
+// in v12 test suite
+
+// shows up in snapshots of json-mode logs
+require('os').hostname = () => 'mock-host'
+
+import * as Lo from 'lodash'
+import * as Logger from '../src'
+import { createMockOutput, MockOutput } from './__helpers'
+
+resetBeforeEachTest(process, 'env')
+resetBeforeEachTest(process.stdout, 'isTTY')
+resetBeforeEachTest(console, 'log')
+
+let log: Logger.RootLogger
+let output: MockOutput
+
+beforeEach(() => {
+  process.env.LOG_PRETTY = 'false'
+  output = createMockOutput()
+  log = Logger.create({ output, pretty: { timeDiff: false } })
+  process.stdout.columns = 200
+})
+
+describe('name', () => {
+  it('becomes the first entry in path', () => {
+    Logger.create({ output, name: 'foo' }).info('bar')
+    expect(output.memory.json[0].path).toEqual(['foo'])
+  })
+  it('defaults to "root"', () => {
+    log.info('bar')
+    expect(output.memory.json[0].path).toEqual(['root'])
+  })
+})
+
+describe('demo', () => {
+  // it('runs a demo with fake data, pretty, all levels active', () => {
+  //   if (!process.version.match(/^v12/)) return
+  //   output.captureConsoleLog()
+  //   log.settings({ pretty: { color: false } })
+  //   Logger.demo(log)
+  //   expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  // })
+  it.todo('runs automatically when LOG_DEMO=true')
+})
+
+describe('output', () => {
+  it('defaults to stdout for all levels', () => {
+    const write = process.stdout.write
+    ;(process.stdout.write as any) = output.write
+    Logger.create({ pretty: false }).fatal('foo')
+    process.stdout.write = write
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+})
+
+describe('level', () => {
+  describe('precedence', () => {
+    it('considers instance time config first', () => {
+      process.env.NODE_ENV = 'production'
+      process.env.LOG_LEVEL = 'fatal'
+      const l = Logger.create({ level: 'fatal' })
+      l.settings({ level: 'trace' })
+      expect(l.settings.level).toEqual('trace')
+    })
+
+    it('then considers construction time config', () => {
+      process.env.NODE_ENV = 'production'
+      process.env.LOG_LEVEL = 'fatal'
+      const l = Logger.create({ level: 'trace' })
+      expect(l.settings.level).toEqual('trace')
+    })
+
+    it('then considers LOG_LEVEL env var', () => {
+      process.env.NODE_ENV = 'production'
+      process.env.LOG_LEVEL = 'trace'
+      const l = Logger.create()
+      expect(l.settings.level).toEqual('trace')
+    })
+
+    it('then considers NODE_ENV=production', () => {
+      process.env.NODE_ENV = 'production'
+      const l = Logger.create()
+      expect(l.settings.level).toEqual('info')
+    })
+
+    it('then defaults to debug', () => {
+      const l = Logger.create()
+      expect(l.settings.level).toEqual('debug')
+    })
+  })
+
+  it('logs below set level are not output', () => {
+    log.settings({ level: 'warn' }).info('foo')
+    expect(output.memory.jsonOrRaw).toEqual([])
+  })
+
+  it('LOG_LEVEL env var config is treated case insensitive', () => {
+    process.env.NODE_ENV = 'production'
+    process.env.LOG_LEVEL = 'TRACE'
+    const l = Logger.create()
+    expect(l.settings.level).toEqual('trace')
+  })
+
+  it('LOG_LEVEL env var config when invalid triggers thrown readable error', () => {
+    process.env.LOG_LEVEL = 'ttrace'
+    expect(() => Logger.create()).toThrowErrorMatchingInlineSnapshot(
+      `"Could not parse environment variable LOG_LEVEL into LogLevel. The environment variable was: ttrace. A valid environment variable must be like: fatal, error, warn, info, debug, trace"`
+    )
+  })
+})
+
+describe('.<level> log methods', () => {
+  it('accept an event name and optional context', () => {
+    log.info('hi', { user: { id: 1 } })
+    log.info('bye')
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+
+  it('one for each log level', () => {
+    log.settings({ level: 'trace' })
+    log.fatal('hi')
+    log.error('hi')
+    log.warn('hi')
+    log.info('hi')
+    log.debug('hi')
+    log.trace('hi')
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+})
+
+describe('.addToContext', () => {
+  it('pins context for all subsequent logs from the logger', () => {
+    log.addToContext({ user: { id: 1 } })
+    log.info('hi')
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+
+  it('can be called multiple times, merging deeply', () => {
+    log.addToContext({ user: { id: 1 } })
+    log.addToContext({ user: { name: 'Jill' } })
+    log.info('hi')
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+
+  it('gets deeply merged with local context', () => {
+    log.addToContext({ user: { id: 1 } })
+    log.info('hi', { user: { name: 'Jill' } })
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+
+  it('local context takes prescedence over pinned context', () => {
+    log.addToContext({ user: { id: 1 } })
+    log.info('hi', { user: { id: 2 } })
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+})
+
+describe('.child', () => {
+  it('creates a sub logger', () => {
+    log.child('tim').info('hi')
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+
+  it('log output includes path field showing the logger namespacing', () => {
+    log.child('b').child('c').child('d').info('foo')
+    expect(output.memory.json[0].path).toEqual(['root', 'b', 'c', 'd'])
+  })
+
+  it('inherits context from parent', () => {
+    log.addToContext({ foo: 'bar' }).child('tim').info('hi')
+    expect(output.memory.json[0].context).toEqual({ foo: 'bar' })
+  })
+
+  it('at log time reflects the current state of parent context', () => {
+    const b = log.child('b')
+    log.addToContext({ foo: 'bar' })
+    b.info('lop')
+    expect(output.memory.json[0].context).toEqual({ foo: 'bar' })
+  })
+
+  it('at log time reflects the current state of parent context even from further up the chain', () => {
+    const b = log.child('b')
+    const c = b.child('c')
+    const d = c.child('d')
+    log.addToContext({ foo: 'bar' })
+    d.info('lop')
+    expect(output.memory.json[0].context).toEqual({ foo: 'bar' })
+  })
+
+  it('inherits level from parent', () => {
+    expect(log.settings.level).toBe('debug')
+    log.settings({ level: 'trace' }).child('tim').trace('hi')
+    // The fact that we get output for trace log from child means it honored the
+    // setLevel.
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+
+  it('reacts to level changes in root logger', () => {
+    const b = log.child('b')
+    log.settings({ level: 'trace' })
+    b.trace('foo')
+    // The fact that we get output for trace log from child means it honored the
+    // setLevel.
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+
+  it('is unable to change context of parent', () => {
+    log.child('b').addToContext({ foo: 'bar' })
+    log.info('qux')
+    expect(output.memory.json[0].context).toEqual({})
+  })
+
+  it('is unable to change context of siblings', () => {
+    const b1 = log.child('b1').addToContext({ from: 'b1' })
+    const b2 = log.child('b2').addToContext({ from: 'b2' })
+    const b3 = log.child('b3').addToContext({ from: 'b3' })
+    log.addToContext({ foo: 'bar' })
+    b1.info('foo')
+    b2.info('foo')
+    b3.info('foo')
+    // All should inherit the root context
+    expect(output.memory.jsonOrRaw).toMatchSnapshot()
+  })
+
+  it('cannot affect level', () => {
+    expect((log.child('b') as any).setLevel).toBeUndefined()
+  })
+})
+
+//
+// Helpers
+//
+
+/**
+ * Restore the key on given object before each test. Useful for permiting tests
+ * to modify the environment and so on.
+ */
+function resetBeforeEachTest(object: any, key: string) {
+  const orig = object[key]
+  beforeEach(() => {
+    if (typeof orig === 'object') {
+      object[key] = Lo.cloneDeep(orig)
+    } else {
+      object[key] = orig
+    }
+  })
+}
